downloads migrated

+++ app/src/main/java/de/escoand/readdaily/database/dao/DownloadDao.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017 escoand.
+ * Copyright (c) 2018 escoand.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -47,6 +47,6 @@ public interface DownloadDao {
     Download findBySubscription(String subscription);
 
     @Query("SELECT * FROM download WHERE downloadId = :downloadId")
-    Download findByDownloadId(int downloadId);
+    Download findByDownloadId(long downloadId);
 
 }
